@@ -4,8 +4,10 @@ export default {
     display: 'inline-block',
     padding: '5px'
   },
+  phoneIcon : {
+    margin: '0 auto'
+  },
   deviceBoxContent : {
-    background: 'green',
     height: '100px',
     padding: '5px'
   },
