@@ -6,17 +6,33 @@ import * as actions from '../redux/actions/index';
 import adb from '../utils/adb';
 
 class HomePage extends Component {
-  updateDevices() {
-    adb.listDevices().then(devices => this.props.actions.updateDevices(devices)).catch(err => console.log(err));
+  listDevices() {
+    const vm = this;
+    adb.listDevices().then(devices => {
+      vm.props.actions.updateDevices(devices);
+      devices.forEach((eachDevice) => {
+        adb.getDeviceProperties(eachDevice.serial).then((deviceProperties) => {
+          vm.props.actions.updateDeviceProperties(eachDevice.serial, deviceProperties);
+        }).catch(err => {
+          console.log(err);
+          vm.props.actions.updateDeviceProperties(eachDevice.serial, {});
+        });
+      });
+    }).catch(err => console.log(err));
   }
-  componentDidMount(){
-    this.updateDevices();
+  onDeviceSelect(device) {
+    const vm = this;
+    vm.props.actions.selectDevice(device.serial);
+  }
+  componentDidMount() {
+    this.listDevices();
   }
   render() {
+    const vm = this;
     return (
       <div>
         {/*<Test content={this.props.state.sample.payload} onClick={this.props.actions.sampleAction}/>*/}
-        <DeviceList devices={this.props.state.devices.devicesList}></DeviceList>
+        <DeviceList onDeviceClick={(device) => vm.onDeviceSelect(device)} devices={vm.props.state.devices.devicesList}></DeviceList>
       </div>
     );
   }
